+++ lib/src/tabbed_view.dart
@@ -48,7 +48,8 @@ class TabbedView extends StatefulWidget {
       this.contentClip = true,
       this.closeButtonTooltip,
       this.tabsAreaButtonsBuilder,
-      this.draggableTabBuilder});
+      this.draggableTabBuilder,
+      this.tabsAreaVisible = true});
 
   final TabbedViewController controller;
   final bool contentClip;
@@ -61,6 +62,7 @@ class TabbedView extends StatefulWidget {
   final String? closeButtonTooltip;
   final TabsAreaButtonsBuilder? tabsAreaButtonsBuilder;
   final DraggableTabBuilder? draggableTabBuilder;
+  final bool tabsAreaVisible;
 
   @override
   State<StatefulWidget> createState() => _TabbedViewState();
@@ -106,11 +108,15 @@ class _TabbedViewState extends State<TabbedView> {
         menuItems: _menuItems);
 
     Widget tabArea = TabsArea(data: data);
-    ContentArea contentArea = ContentArea(data: data);
-    return CustomMultiChildLayout(children: [
-      LayoutId(id: 1, child: tabArea),
-      LayoutId(id: 2, child: contentArea)
-    ], delegate: _TabbedViewLayout());
+    ContentArea contentArea =
+        ContentArea(data: data, tabsAreaVisible: widget.tabsAreaVisible);
+    List<LayoutId> children = [];
+    if (widget.tabsAreaVisible) {
+      children.add(LayoutId(id: 1, child: tabArea));
+    }
+    children.add(LayoutId(id: 2, child: contentArea));
+    return CustomMultiChildLayout(
+        children: children, delegate: _TabbedViewLayout());
   }
 
   /// Set a new menu items.
@@ -148,14 +154,17 @@ class _TabbedViewState extends State<TabbedView> {
 class _TabbedViewLayout extends MultiChildLayoutDelegate {
   @override
   void performLayout(Size size) {
-    Size childSize = layoutChild(
-        1,
-        BoxConstraints(
-            minWidth: size.width,
-            maxWidth: size.width,
-            minHeight: 0,
-            maxHeight: size.height));
-    positionChild(1, Offset.zero);
+    Size childSize = Size.zero;
+    if (hasChild(1)) {
+      childSize = layoutChild(
+          1,
+          BoxConstraints(
+              minWidth: size.width,
+              maxWidth: size.width,
+              minHeight: 0,
+              maxHeight: size.height));
+      positionChild(1, Offset.zero);
+    }
     double height = math.max(0, size.height - childSize.height);
     layoutChild(2, BoxConstraints.tightFor(width: size.width, height: height));
     positionChild(2, Offset(0, childSize.height));