+++ lib/src/theme/default_themes/dark_theme.dart
@@ -113,6 +113,7 @@ class DarkTheme {
       {required Color selectedTabColor}) {
     return ContentAreaThemeData(
         decoration: BoxDecoration(color: selectedTabColor),
+        decorationNoTabsArea: BoxDecoration(color: selectedTabColor),
         padding: EdgeInsets.all(8));
   }