+++ lib/src/theme/tab_status_theme_data.dart
@@ -44,4 +44,39 @@ class TabStatusThemeData {
   BoxDecoration? normalButtonBackground;
   BoxDecoration? hoverButtonBackground;
   BoxDecoration? disabledButtonBackground;
+
+  @override
+  bool operator ==(Object other) =>
+      identical(this, other) ||
+      other is TabStatusThemeData &&
+          runtimeType == other.runtimeType &&
+          padding == other.padding &&
+          paddingWithoutButton == other.paddingWithoutButton &&
+          margin == other.margin &&
+          decoration == other.decoration &&
+          innerBottomBorder == other.innerBottomBorder &&
+          innerTopBorder == other.innerTopBorder &&
+          fontColor == other.fontColor &&
+          normalButtonColor == other.normalButtonColor &&
+          hoverButtonColor == other.hoverButtonColor &&
+          disabledButtonColor == other.disabledButtonColor &&
+          normalButtonBackground == other.normalButtonBackground &&
+          hoverButtonBackground == other.hoverButtonBackground &&
+          disabledButtonBackground == other.disabledButtonBackground;
+
+  @override
+  int get hashCode =>
+      padding.hashCode ^
+      paddingWithoutButton.hashCode ^
+      margin.hashCode ^
+      decoration.hashCode ^
+      innerBottomBorder.hashCode ^
+      innerTopBorder.hashCode ^
+      fontColor.hashCode ^
+      normalButtonColor.hashCode ^
+      hoverButtonColor.hashCode ^
+      disabledButtonColor.hashCode ^
+      normalButtonBackground.hashCode ^
+      hoverButtonBackground.hashCode ^
+      disabledButtonBackground.hashCode;
 }